+++ src/services/find_me_service.test.ts
@@ -1,5 +1,5 @@
 import { HAP } from "homebridge";
-import * as HapJs from "hap-nodejs";
+import * as HapJs from "@homebridge/hap-nodejs";
 import { Subject } from "rxjs";
 import { FindMeService } from "./find_me_service";
 import {