+++ src/utils/ensure_name.ts
@@ -7,7 +7,7 @@ export function ensureName(hap: HAP, service: Service, name: string) {
     name.replaceAll(" ", "_"),
   ].join("-");
   service.addOptionalCharacteristic(hap.Characteristic.ConfiguredName);
-  if (!hap.HAPStorage.storage().getItemSync(key)) {
+  if (!hap.HAPStorage.storage().getItem(key)) {
     service.setCharacteristic(hap.Characteristic.ConfiguredName, name);
   }
   service